Watch for connection stress, limit new connections
Report if connection is stressed (Connecting, or has available window to
send that is still unsent).

Using this information, hold off on new connection attempts when stress
is higher than the number of worker threads, because we can reasonably
deduce a *local* problem keeping up, and so we shouldn't add more load
to ourselves with a new connection.

Also, report some connection stress stats in the logs.