Some config values make test less flaky, updating
Longer idle timeout has obvious effect of reducing connection failures
due to idle timeout. This idle timeout doesn't seem to be specified
anywhere, we are just doing it. 10s seemed reasonable, but when fighting
against tokio which is not evenly distributing tasks, it gets less
reasonable.

The retry timeout should be small enough that if it maxes out, there are
a few chances at reattempts before the idle timeout. Setting it to /4 of
the idle timeout means we should get at least 3 attempts (the 4th will
race with the idle).

Was seeing regular connection failures, even with the stress limiter. By
bumping up to 5 attempts, it seemed to effectively disappear. I don't
have a strong model for how often SYNs are dropped in this test. To get
an estimate of the allowed failure rate, making a silly-ish assumption
that each attempt has an independently random chance of failing, then to
have the test pass 99.9% of the time, we need each connection attempt to
succeed at least 75% of the time.