Mock Socket & test that close() waits for ACK
Use the new mock socket to:
- remove chance of real dropped packets on UDP during test
- allow parallel tests without fiddling with UDP ports
- convenience method to launch connected peers
- test what happens when network goes down

Add test that close() waits for writes to ACK and eventually returns a
timeout, by disabling the network after connecting to a peer. Closing
the connection will time out, because the sender is waiting for ACKs
that the recipient will never send. (Because the recipient never
recieves the data)

Why does this add the tokio test-util and start_paused = true?

We don't want the test to actually wait the timeout length while
running, so we use tokio's pause feature to skip over the timeout
without interrupting anything else that would happen in real-time:
https://docs.rs/tokio/latest/tokio/time/fn.pause.html