fix: ack sent packets using all received packets
The ack_nr on received packets was previously ignored in everything but
State packets. In some scenarios, like with dropped packets or different
utp implementations, that would cause the sender to wait indefinitely
for the recipient to ACK state. But the recipient had already acked with
the FIN (been ignored), and disconnected. Then the sender was waiting
the full connection timeout to realize it.

Even worse, if more than just the final outbound FIN was acknowledged in
the inbound FIN, the sender was considering the data transfer
incomplete.

With this commit:
- When the recipient acks the final FIN with a FIN, the sender doesn't
  get stuck waiting for the timeout, and returns from close()
  immediately.
- When the recipient acks some data with the FIN, the sender now returns
  from close() with an OK instead of Err.

Also added a test of acknowledging packets.