test: run 2k concurrent transfers with 1MB of data
Now we can reliably handle the launching of many more streams, so let's
really stress-test the library.

This has helped us identify and resolve a number of bugs, in local
testing. Now that they seem to be fixed, we can merge to master to
prevent regressions.

Bonus cleanups on @njgheorghita 's original change:
- Change the high end of the range to equal the number of concurrent
  streams
- Log showing the transfer speed in socket test
- Run an optimized build of utp when testing, for more performance (and
  interesting results) in the high-concurrency tests
- Return from the test as soon as an error is found, cancelling all
  in-flight transfers