refactor: more general options for testing links
The only option currently is to manually control the link between
transferring peers.

We want to add a test that drops a specific packet, where manual control
is insufficient, so we want to expand the concept of managing whether
the link is up or down. Now we can make arbitrary logic to allow packets
to pass or drop.