fix: receive all data during SYN->STATE drop
The bug occurs when the STATE packet in response to the SYN is dropped,
and STATE-sender starts transmitting data before receiving the SYN
retransmission. The repeated SYN is returned with another STATE packet,
but that STATE packet sets the next sequence number to be after the
already sent data. So all previously sent data is silently dropped by
the recipient, corrupting the transmission.

The fix is to copy the first STATE response and retransmit it in
response to any duplicate SYN packets received.

A test case that replicates the bug is attached.